memunreachable: prevent forking when main thread times out
The main thread waits up to 30 seconds for the collection thread to
pause all the other threads.  If the collection thread has not
finished pausing after 30 seconds the main thread will release
the malloc locks.  That could allow an unpaused thread to take a
malloc lock and then get paused while holding it.  When the collection
thread tries to fork it will take all the malloc locks again,
deadlocking with the paused thread that is holding a malloc lock.

Switch from a semaphore to an atomic state machine to coordinate
between the main thread adn the collection thread.  If the main
thread stops waiting it will atomically transition the state to
ABORT, which will prevent the collection thread from forking.

Bug: 231273975
Test: memunreachable_test, memunreachable_unit_test
Change-Id: Ia37b5612dd0434f0231e2aeb159a8e9e5e6afd5d
Merged-In: Ia37b5612dd0434f0231e2aeb159a8e9e5e6afd5d