Remove mutex from AtomicState
AtomicState can't use a mutex, if the main thread is pthread stopped
while it is holding the mutex inside AtomicState.wait_for_either_of
it could block the mutex forever, deadlocking with the collection
thread that needs the mutex before it can unpause the main thread.
Switch to using std::atomic with a busy loop during wait_for_either_of.

Fixes: 288344946
Test: AtomicState_test
Test: Memunreachable_test
Change-Id: I2ae0705b78e04c4013c9b1735a9c0142a89757bb