Don't treat libmemunreachable's stack as a global root
libmemunreachable's ptracer thread allocates itself a stack.  The stack
is used by the heap walker, and could contain previous pointers that the
heap walker has examined.  When the heap walker examines the stack it
could find the previously examined pointers, causing false negatives.
Skip the libmemunreachable_stack named mapping so that it is not used
as a root.

Fixes flaky memunreachable_test failurse.

Fixes: 292297017
Test: adb shell 'while /data/nativetest64/memunreachable_test/memunreachable_test --gtest_filter=*.stack; do true; done'
Change-Id: Id8f7cd12cd06c9f555143ad2c80a25e029483c1d